fix(example): SEO-law compliance — server page + client BuyDemo child
