+++ apps/web/app/examples/dodo-checkout/page.tsx
@@ -1,30 +1,16 @@
-"use client";
+import { pageMetadata } from "@stack/seo";
+import BuyDemo from "./BuyDemo";
 
-import { useState } from "react";
-import CheckoutModal from "./CheckoutModal";
+// Minimal demo page: server-rendered product card shell; the interactive Buy
+// flow (fetch → CheckoutModal) lives in the BuyDemo client child.
+export const metadata = pageMetadata({
+  title: "White-label Checkout Example",
+  description:
+    "Demo of a white-label Dodo Payments checkout — a product card that opens the hosted checkout inside an on-domain modal.",
+  path: "/examples/dodo-checkout",
+});
 
-// Minimal demo page: product card → POST /api/examples/dodo-checkout → embed
-// the returned checkout url in the white-label CheckoutModal.
 export default function DodoCheckoutExamplePage() {
-  const [checkoutUrl, setCheckoutUrl] = useState<string | null>(null);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  async function buy() {
-    setLoading(true);
-    setError(null);
-    try {
-      const res = await fetch("/api/examples/dodo-checkout", { method: "POST" });
-      if (!res.ok) throw new Error(`Checkout failed (${res.status})`);
-      const { url } = (await res.json()) as { url: string };
-      setCheckoutUrl(url);
-    } catch (e) {
-      setError(e instanceof Error ? e.message : "Checkout failed");
-    } finally {
-      setLoading(false);
-    }
-  }
-
   return (
     <main
       style={{
@@ -48,34 +34,8 @@ export default function DodoCheckoutExamplePage() {
       >
         <h1 style={{ margin: 0, fontSize: 18, fontWeight: 600, color: "#171717" }}>Demo Product</h1>
         <p style={{ margin: "8px 0 24px", fontSize: 14, color: "#525252" }}>$10.00</p>
-        <button
-          type="button"
-          onClick={buy}
-          disabled={loading}
-          style={{
-            width: "100%",
-            padding: "12px 32px",
-            border: "none",
-            borderRadius: 4,
-            background: "#111",
-            color: "#fff",
-            fontSize: 13,
-            cursor: loading ? "wait" : "pointer",
-            opacity: loading ? 0.7 : 1,
-          }}
-        >
-          {loading ? "Preparing checkout…" : "Buy now"}
-        </button>
-        {error && (
-          <p role="alert" style={{ marginTop: 12, fontSize: 12, color: "#b91c1c" }}>
-            {error}
-          </p>
-        )}
+        <BuyDemo />
       </div>
-
-      {checkoutUrl && (
-        <CheckoutModal checkoutUrl={checkoutUrl} onClose={() => setCheckoutUrl(null)} />
-      )}
     </main>
   );
 }
